global: refactore and add new command
* move functions of cli.py to utils.py

* add command to fetch all ids

* add experimental celery job
